Fix "keys changed during iteration" error
This change is necessary for compatibility with Python 3.8, where iterating over a changing dictionary raises and error as opposed the warning produced by previous versions. The solution is just to iterate over a list with the keys of the dictionary instead of a view object. For further info see https://cito.github.io/blog/never-iterate-a-changing-dict/